node: move VMExecution mocked impl to the binary
