+++ node/src/vm/mod.rs
@@ -25,40 +25,3 @@ pub trait VMExecution: Send + Sync + 'static {
     fn finalize(&self, blk: &Block) -> anyhow::Result<()>;
     fn preverify(&self, tx: &Transaction) -> anyhow::Result<()>;
 }
-
-/// Empty Placeholder for VMExecution
-pub struct VMExecutionImpl {}
-
-impl VMExecutionImpl {
-    pub fn new(conf: Config) -> Self {
-        Self {}
-    }
-}
-
-impl VMExecution for VMExecutionImpl {
-    fn execute_state_transition(
-        &self,
-        txs: &[Transaction],
-    ) -> anyhow::Result<()> {
-        Ok(())
-    }
-
-    fn verify_state_transition(
-        &self,
-        txs: &[Transaction],
-    ) -> anyhow::Result<()> {
-        Ok(())
-    }
-
-    fn accept(&self, blk: &Block) -> anyhow::Result<()> {
-        Ok(())
-    }
-
-    fn finalize(&self, blk: &Block) -> anyhow::Result<()> {
-        Ok(())
-    }
-
-    fn preverify(&self, tx: &Transaction) -> anyhow::Result<()> {
-        Ok(())
-    }
-}
